refactor:cleanup secrets_manager and others
