+++ src/britive/access_broker/__init__.py
@@ -1,13 +1,13 @@
 from .brokers import Brokers
 from .pools import Pools
-from .profiles.profiles import Profile
-from .resources.resources import Resources
+from .profiles import Profiles
+from .resources import Resources
 from .response_templates import ResponseTemplates
 
 
 class AccessBroker:
     def __init__(self, britive) -> None:
-        self.profiles = Profile(britive)
+        self.profiles = Profiles(britive)
         self.resources = Resources(britive)
         self.response_templates = ResponseTemplates(britive)
         self.brokers = Brokers(britive)
